test: move issue #160 regression test into the unit suite
The PR check workflow (`.github/workflows/test.yml`) only runs
`./gradlew -i test jacocoTestReport`, which doesn't include the
`integrationTest` source set. Integration tests do run in `cross.yml`,
but only after a 10+ minute cross-compile pipeline — so a regression
that reintroduces the clone/replay behaviour could pass `test.yml`
silently (codex PR #161 review).

Move the loop test into `src/test/java/` so it runs on every PR via
the fast workflow. `test.yml` already starts SurrealDB at
`ws://localhost:8000` before invoking gradle, so the test has a real
WS server to talk to.

Skipped via `Assumptions.assumeTrue(false, …)` when no server is
reachable, so local `./gradlew test` runs without a server stay green.
